+++ src/gcp/iam.ts
@@ -2,7 +2,7 @@ import { resourceManagerOrigin, iamOrigin } from "../api";
 import { logger } from "../logger";
 import { Client } from "../apiv2";
 
-const apiClient = new Client({ urlPrefix: iamOrigin, apiVersion: "v1" });
+const apiClient = new Client({ urlPrefix: iamOrigin(), apiVersion: "v1" });
 
 // IAM Policy
 // https://cloud.google.com/resource-manager/reference/rest/Shared.Types/Policy
@@ -56,7 +56,6 @@ export interface TestIamResult {
 
 /**
  * Creates a new the service account with the given parameters.
- *
  * @param projectId the id of the project where the service account will be created
  * @param accountId the id to use for the account
  * @param description a brief description of the account
@@ -87,7 +86,6 @@ export async function createServiceAccount(
 
 /**
  * Retrieves a service account with the given parameters.
- *
  * @param projectId the id of the project where the service account will be created
  * @param serviceAccountName the name of the service account
  */
@@ -160,7 +158,6 @@ export async function getRole(role: string): Promise<Role> {
 
 /**
  * List permissions not held by an arbitrary resource implementing the IAM APIs.
- *
  * @param origin Resource origin e.g. `https:// iam.googleapis.com`.
  * @param apiVersion API version e.g. `v1`.
  * @param resourceName Resource name e.g. `projects/my-projct/widgets/abc`
@@ -215,7 +212,7 @@ export async function testIamPermissions(
   permissions: string[],
 ): Promise<TestIamResult> {
   return testResourceIamPermissions(
-    resourceManagerOrigin,
+    resourceManagerOrigin(),
     "v1",
     `projects/${projectId}`,
     permissions,