+++ src/test/deploy/functions/runtimes/python/index.spec.ts
@@ -26,14 +26,6 @@ describe("PythonDelegate", () => {
       expect(delegate.getPythonBinary()).to.equal("python3.10");
     });
 
-    it("returns generic python binary given non-recognized python runtime", () => {
-      platformMock.value("darwin");
-      const requestedRuntime = "python308";
-      const delegate = new python.Delegate(PROJECT_ID, SOURCE_DIR, requestedRuntime);
-
-      expect(delegate.getPythonBinary()).to.equal("python");
-    });
-
     it("always returns version-neutral, python.exe on windows", () => {
       platformMock.value("win32");
       const requestedRuntime = "python310";